// Admin-Panel/src/app/@core/models/provider/provider.error.ts
export class ProviderError {
    name_translations: Array<string>;
    details_translations: Array<string>;
    meta: Array<string>;
    images: Array<string>;
    fee: Array<string>;
    is_verified: Array<string>;
    address: Array<string>;
    longitude: Array<string>;
    latitude: Array<string>;
    categories: Array<string>;
    subcategories: Array<string>;
    availability: Array<string>;
    plan: Array<string>;

    // user related
    email: Array<string>;
    mobile_number: Array<string>;
    password: Array<string>;
}